requirements.txt
certifi==2024.2.2
chardet==4.0.0
idna==2.10
mysql-connector==2.2.9
pyodbc==5.1.0
requests==2.25.1
urllib3==1.26.18
